requirements.txt
fastapi
uvicorn
opencv-python
numpy
scipy
pillow
python-multipart